Cleanup and refactoring.
